requirements.txt
numpy
matplotlib
Pillow
streamlit
scikit-learn